Struct ckb_testtool::context::Context

pub struct Context {
    pub cells: HashMap<OutPoint, (CellOutput, Bytes)>,
    pub headers: HashMap<Byte32, HeaderView>,
    pub epoches: HashMap<Byte32, EpochExt>,
    pub cells_by_data_hash: HashMap<Byte32, OutPoint>,
}

Verification Context

Fields
cells: HashMap<OutPoint, (CellOutput, Bytes)>headers: HashMap<Byte32, HeaderView>epoches: HashMap<Byte32, EpochExt>cells_by_data_hash: HashMap<Byte32, OutPoint>
Methods

impl Context

pub fn deploy_contract(&mut self, data: Bytes) -> OutPoint

Deprecated since 0.1.1:
Please use the deploy_cell function instead

pub fn deploy_cell(&mut self, data: Bytes) -> OutPoint

Deploy a cell return the out-point of the cell

pub fn get_contract_out_point(&self, data_hash: &Byte32) -> Option<OutPoint>

Deprecated since 0.1.1:
Please use the get_cell_by_data_hash function instead

pub fn get_cell_by_data_hash(&self, data_hash: &Byte32) -> Option<OutPoint>

Get the out-point of a cell by data_hash the cell must has deployed to this context

pub fn create_cell(&mut self, cell: CellOutput, data: Bytes) -> OutPoint

Create a cell with data return the out-point

pub fn create_cell_with_out_point(
&mut self,
out_point: OutPoint,
cell: CellOutput,
data: Bytes
)

Create cell with specified out-point and cell data

pub fn insert_cell(
&mut self,
out_point: OutPoint,
cell: CellOutput,
data: Bytes
)

Deprecated since 0.1.1:
Please use the create_cell_with_out_point function instead

pub fn get_cell(&self, out_point: &OutPoint) -> Option<(CellOutput, Bytes)>

Get cell output and data by out-point

pub fn build_script(
&mut self,
out_point: &OutPoint,
args: Bytes
) -> Option<Script>

Build script with out_point and args return none if the out-point is not exist

pub fn complete_tx(&mut self, tx: TransactionView) -> TransactionView

Complete cell deps for a transaction this function searches context cells; generate cell dep for referenced scripts.

pub fn verify_tx(
&self,
tx: &TransactionView,
max_cycles: u64
) -> Result<Cycle, CKBError>

Verify the transaction in CKB-VM